Potential fix for code scanning alert no. 15: Multiplication result converted to larger type

+++ Library/OcPngLib/lodepng.c
@@ -4340,7 +4340,7 @@ static void Adam7_getpassvalues(unsigned passw[7], unsigned passh[7], size_t fil
     filter_passstart[i + 1] = filter_passstart[i]
                             + ((passw[i] && passh[i]) ? passh[i] * (1u + (passw[i] * bpp + 7u) / 8u) : 0);
     /*bits padded if needed to fill full byte at end of each scanline*/
-    padded_passstart[i + 1] = padded_passstart[i] + passh[i] * ((passw[i] * bpp + 7u) / 8u);
+    padded_passstart[i + 1] = padded_passstart[i] + (size_t)passh[i] * ((passw[i] * bpp + 7u) / 8u);
     /*only padded at end of reduced image*/
     passstart[i + 1] = passstart[i] + (passh[i] * passw[i] * bpp + 7u) / 8u;
   }
